requirements.txt
fastapi
uvicorn[standard]
easyocr
torch
transformers
pillow
python-multipart
numpy
opencv-python-headless
torchvision